extract theme settings section
